+++ src/utils/index.ts
@@ -76,3 +76,4 @@ export * from "./CLIUtils";
 export * from "./BNUtils";
 export * from "./CCTPUtils";
 export * from "./RetryUtils";
+export * from "./BinanceUtils";
